Centralize template selector expansion (#82/#61, subsuming #80)
- Add _errors.py, _helpers.py, _symbols.py, _axes.py, _constraints.py
- Add _templates.py: unified parse_selector/expand_selector/render_selector API
  supporting wildcard, pinned, and mixed selectors
- Slim specs.py: remove local template primitives and constraint block,
  wire expanders to _templates.py API
- Wire operator and coord_shift apply_to through expand_apply_to
- Support pinned keys in initial_state (e.g. X[age, vax, loc, imm=X0])
- Update tests: import ConstraintRule/_normalize_constraints from _constraints
